feat: Add Nix support
This adds the files necessary for Nix users to easily use fex.

To put it simply, the `flake.nix` file currently defines what package
outputs are available, and for which systems. I set it as the default
package for convenience.

The `default.nix` file is the actual definition of the `fex` package,
where the package name, version, dependencies, and other information
live.