use Synchronized Output using CSI
more details of what it is is present here:
https://gist.github.com/christianparpart/d8a62cc1ab659194337d73e399004036

on iTerm the way to do this is with the DCS "ESC P = 1 s ESC \" but this
causes garbage output on the macOS Terminal and afaik there is not way
to test for compatibility

the used CSI method, i.e. "ESC [ ? 2026 h" is to being adopted by most
terminals. This fails silently if it is not supported (including the
macOS Terminal).

DECRQM does not appear to be implemented on the macOS terminal so
compatibility cannot be queried using "ESC [ ? 2026 $ p" nevertheless
this commits includes the code to do all of it.